Update picam.py
Generalized ProcessData to work with multiple ROIs. Resulting data is in a list of 3-D numpy arrays. List elements correspond to ROI#.
Disabled the live display because multi-threading seems to be broken. I am losing the camera handle in the new thread. I will work on this, but for now it is advised not to use the live view.
Added a function (SetROIs) to demonstrate how to add multiple ROIs.